🚔 Replace processor tests with integration tests using Gradle TestKit
Replaced existing KSP processor tests with Gradle TestKit-based
integration tests. Removed `CapturingProcessor` and related classes,
simplifying test structure and ensuring tests operate in a real Gradle
KSP environment.
