+++ gradle/libs.versions.toml
@@ -4,7 +4,6 @@ kotlin = "2.0.21"
 kotlinx-html = "0.11.0"
 junit = "5.10.1"
 junit-platform = "1.10.1"
-kotlin-compile-testing = "1.5.0"
 
 ksp = "2.0.21-1.0.27"
 maven-publish = "0.35.0"
@@ -18,7 +17,6 @@ kotlinx-html = { group = "org.jetbrains.kotlinx", name = "kotlinx-html-jvm", ver
 junit-jupiter-api = { group = "org.junit.jupiter", name = "junit-jupiter-api", version.ref = "junit" }
 junit-jupiter-engine = { group = "org.junit.jupiter", name = "junit-jupiter-engine", version.ref = "junit" }
 junit-platform-launcher = { group = "org.junit.platform", name = "junit-platform-launcher", version.ref = "junit-platform" }
-kotlin-compile-testing-ksp = { group = "com.github.tschuchortdev", name = "kotlin-compile-testing-ksp", version.ref = "kotlin-compile-testing" }
 
 [plugins]
 kotlin-jvm = { id = "org.jetbrains.kotlin.jvm", version.ref = "kotlin" }
