+++ techdebt-processor/build.gradle.kts
@@ -1,6 +1,7 @@
 plugins {
     alias(libs.plugins.kotlin.jvm)
     alias(libs.plugins.maven.publish)
+    `java-gradle-plugin`
 }
 
 group = "io.github.igorescodro"
@@ -18,7 +19,7 @@ dependencies {
     testImplementation(libs.junit.jupiter.api)
     testRuntimeOnly(libs.junit.jupiter.engine)
     testRuntimeOnly(libs.junit.platform.launcher)
-    testImplementation(libs.kotlin.compile.testing.ksp)
+    testImplementation(gradleTestKit())
 }
 
 tasks.test {
